Order service completed with docker mysql container

+++ common-util/src/main/java/com/sahu/sagachoreography/common/dto/PaymentRequestDTO.java
@@ -2,11 +2,13 @@ package com.sahu.sagachoreography.common.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record PaymentRequestDTO (
     Long orderId,
     Long userId,
-    Double amount
+    BigDecimal amount
 ){
 
 }
